redsocks: Fix dnsu2t dns relay when there is no TFO cookie cached
This adds a patch over dnsu2t:

The current implementation uses TFO when resending DNS request over TCP.
This works well when there is a cookie cached as the data will be send
with the SYN packet.

The problem is at the first connection request. When that happens,
(first sendto()) the client will send a SYN packet with TFO option being
it a TFO request but given that we are using a non-blocking socket,
sendto will not send the data too and return errno EINPROGRESS[1]. It
assumes that a write is retried after that cookie exchange is done[2].

This looks like an ignorable issue as it theoretically only makes the
first request fail but it gets bad when the middleware strips TFO
options. See for example Juniper routers[3]. In this case the client
will never get a cookie back and will never send data either as every
sendto will just send a SYN with TFO request which ends up as a simple
SYN on the server side. The server will just proceed with a classic 3WHS
but the client will never pass the request afterwards.

This patch saves the DNS request when a cookie request is sent and
resends the data when the socket becomes writable (in the appropiate
callback function).

[1] https://www.spinics.net/lists/netdev/msg205324.html
[2] https://www.spinics.net/lists/linux-man/msg11329.html
[3] https://www.juniper.net/documentation/en_US/junos/topics/task/configuration/tfo-configuring.html#jd0e18

Signed-off-by: Andrei Gherzan <andrei@resin.io>